# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: WashBuddy API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /providers/{providerId}/locations:
    get:
      operationId: listLocations
      tags: [locations]
      summary: List provider locations
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      responses:
        "200":
          description: Location list
          content:
            application/json:
              schema:
                type: object
                properties:
                  locations:
                    type: array
                    items:
                      $ref: "#/components/schemas/Location"
                required: [locations]
    post:
      operationId: createLocation
      tags: [locations]
      summary: Create location
      parameters:
        - $ref: "#/components/parameters/ProviderId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    $ref: "#/components/schemas/Location"
                required: [location]
  /providers/{providerId}/locations/{locationId}:
    patch:
      operationId: updateLocation
      tags: [locations]
      summary: Update location
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/LocationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLocationInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  location:
                    $ref: "#/components/schemas/Location"
                required: [location]
  /providers/{providerId}/locations/{locationId}/hours:
    put:
      operationId: setOperatingHours
      tags: [locations]
      summary: Set operating hours
      parameters:
        - $ref: "#/components/parameters/ProviderId"
        - $ref: "#/components/parameters/LocationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                windows:
                  type: array
                  items:
                    $ref: "#/components/schemas/OperatingWindowInput"
              required: [windows]
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  windows:
                    type: array
                    items:
                      $ref: "#/components/schemas/OperatingWindow"
                required: [windows]
  /locations/search:
    get:
      operationId: searchLocations
      tags: [locations]
      summary: Search visible locations
      parameters:
        - name: lat
          in: query
          schema:
            type: number
        - name: lng
          in: query
          schema:
            type: number
        - name: radiusMi
          in: query
          schema:
            type: number
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Location search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  locations:
                    type: array
                    items:
                      $ref: "#/components/schemas/LocationSearchResult"
                required: [locations]
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List user notifications
      parameters:
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                required: [success]
...
components:
  parameters:
    ProviderId:
      name: providerId
      in: path
      required: true
      schema:
        type: string
    LocationId:
      name: locationId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
...
    Provider:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        legalName:
          type: ["string", "null"]
        slug:
          type: string
        contactEmail:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        billingMode:
          type: string
        payoutReady:
          type: boolean
        createdAt:
          type: string
      required: [id, name, slug, billingMode, payoutReady]
...
    Location:
      type: object
      properties:
        id:
          type: string
        providerId:
          type: string
        name:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        timezone:
          type: string
        isVisible:
          type: boolean
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        responseSlaFutureMins:
          type: integer
        responseSlaUnder1hMins:
          type: integer
        provider:
          $ref: "#/components/schemas/Provider"
        operatingWindows:
          type: array
          items:
            $ref: "#/components/schemas/OperatingWindow"
      required: [id, providerId, name, addressLine1, city, stateCode, postalCode, timezone, isVisible]
    CreateLocationInput:
      type: object
      properties:
        name:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        countryCode:
          type: string
        timezone:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        isVisible:
          type: boolean
      required: [name, addressLine1, city, stateCode, postalCode, timezone]
    UpdateLocationInput:
      type: object
      properties:
        name:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        timezone:
          type: string
        isVisible:
          type: boolean
        latitude:
          type: number
        longitude:
          type: number
    LocationSearchResult:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        stateCode:
          type: string
        postalCode:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        provider:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
          required: [id, name]
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceSummary"
        operatingWindows:
          type: array
          items:
            $ref: "#/components/schemas/OperatingWindowSummary"
      required: [id, name, addressLine1, city, stateCode, postalCode]
    OperatingWindowSummary:
      type: object
      properties:
        dayOfWeek:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
      required: [dayOfWeek, openTime, closeTime]
    OperatingWindow:
      type: object
      properties:
        id:
          type: string
        dayOfWeek:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
      required: [id, dayOfWeek, openTime, closeTime]
    OperatingWindowInput:
      type: object
      properties:
        dayOfWeek:
          type: integer
        openTime:
          type: string
        closeTime:
          type: string
      required: [dayOfWeek, openTime, closeTime]
...
    ServiceSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        basePriceMinor:
          type: integer
        durationMins:
          type: integer
      required: [id, name, basePriceMinor, durationMins]
...
    NotificationItem:
      type: object
      properties:
        id:
          type: string
        subject:
          type: ["string", "null"]
        body:
          type: string
        actionUrl:
          type: ["string", "null"]
        metadata:
          type: object
        readAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, body, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/NotificationItem"
        hasMore:
          type: boolean
        nextCursor:
          type: ["string", "null"]
      required: [notifications, hasMore]